Don't attempt to create the pubsub topic
The user doesn't have IAM permissions to check the existence of topics or permission to create them with this setup. Just grab the topic and move along - it will fail downstream if the permissions are not setup or if the topic does not exist, but you guide people through the setup.