Fix bumps in first-time setup (#4)
* Don't prompt to delete pubsub topic on cleanup

* Enable required services as part of setup

Without this, other commands fail (e.g. you can't talk to bg without enabling the bq API first)

* Pass region to service IAM policy binding

Without this, the command isn't automated (it prompts for input)

* Print topicName, not topic

topic is a struct and will always be nil at this point

* Run setup before building the image

Cloud build needs to be enabled first. It doesn't do us a ton of good to build the container before setup, and it doesn't make sense logically to the user.

* Strip symbols too

Smaller binary

* Don't attempt to create the pubsub topic

The user doesn't have IAM permissions to check the existence of topics or permission to create them with this setup. Just grab the topic and move along - it will fail downstream if the permissions are not setup or if the topic does not exist, but you guide people through the setup.

* Use SERVICE_REGION in URL, platform managed